getting a test attempt by its id now returns the score from interview bot

+++ Models/IMochaDTOs/IMochaCandidateTestReport.cs
@@ -8,7 +8,7 @@ public class CandidateTestReport
     public string candidateName { get; set; } = "";
     public DateTime attemptedOn { get; set; }
     public string status { get; set; } = "";
-    public double score { get; set; }
+    public decimal score { get; set; }
     public double candidatePoints { get; set; }
     public double totalTestPoints { get; set; }
     public double scorePercentage { get; set; }
